p1-t8: typed IPC layer via tauri-specta — spike PASSED
TECH_ARCHITECTURE.md §2.3. Spike result: PASS. All 10 commands
annotated and bindings generated correctly on the first real attempt
once the dependency-resolution issue below was solved — including
both target-hard cases (render_audiogram's RenderJobDto with its
Vec<f32> peaks field, and analyze_spectrum's SpectrumResult).

## The actual spike risk: pinning a working tauri-specta rc

`cargo add specta` / `tauri-specta` with no version resolves to
ancient stable releases (v1.0.5 / v1.0.2) that don't know about Tauri
2 or the derive/typescript feature flags at all. The 2.x line is still
all pre-release rc builds. Bisected by hand:
- rc.15/rc.19 pull in tauri-specta's OWN pinned pre-release tauri
  (=2.0.0-rc.0 / =2.0.0-rc.10) — conflicts with our stable tauri 2.7.0.
- rc.21+ (up to the current rc.25) pull in a specta version that uses
  `std::fmt::from_fn`, an unstable stdlib feature not available on
  stable rustc 1.89.0 — hard compile error, not a code issue on our
  side.
- rc.20 is the sweet spot: last version before that unstable-feature
  regression, first version that accepts stable tauri ^2. Pinned exactly
  (specta, tauri-specta, specta-typescript all at 2.0.0-rc.20/0.0.7).

## Real compile-time findings once rc.20 built

specta's TypeScript exporter rejects u64/i64/usize/isize by default
(BigIntForbidden) since JS `number` can't represent 64-bit ints
precisely. Three pre-existing fields were wider than they needed to be
for their actual value range — narrowed, not worked around:
- ModelSpec/ModelInfo.size_mb: u64 -> u32 (largest catalogued model is
  ~3.1 GB in MB; u32 tops out around 4.29 billion)
- Segment.id: usize -> u32 (a transcript's segment index)
- SpectrumResult.n_buckets/n_bands: usize -> u32
Each change required casting back to usize/u64 at the couple of call
sites that use these as real array indices or byte-count math
(model_repo.rs's size_bytes computation, render/mod.rs's fft_n_buckets,
runner.rs's segment id from .enumerate()) — the local computations stay
wide, only the values that cross the IPC boundary got narrowed.

Also gave the generated bindings.gen.ts a custom `// @ts-nocheck`
header (via specta_typescript::Typescript::header()) — it ships unused
event-plumbing boilerplate (TAURI_CHANNEL, __makeEvents__) since we
don't use tauri-specta's event-mapping feature, which trips
noUnusedLocals; excluded from eslint entirely too, same reasoning.
Committing bindings.gen.ts (not gitignoring it) since it's needed to
resolve `./bindings.gen` imports without a Rust build in the loop —
matches T18's CI split (frontend job typechecks independently of the
rust job).

## New files

- core/errors.ts: AppError {kind, message, detail?, retryable} +
  toAppError() normalizer. Rust commands return Result<T, String>
  (no structured error type crossing IPC yet), so classification is a
  substring heuristic on that string per the plan's own mapping
  ("ffmpeg"->ffmpeg-missing, "whisper"|"model"->whisper-failed, else
  unknown) plus a couple of obvious extras (decode/cancel/io).
- core/ipc/client.ts: the only file allowed to import bindings.gen.ts.
  Wraps every generated command, unwrapping the Result<T,E> tag union
  or catching an IPC-level throw, always rejecting with one AppError
  shape instead of a raw string or bare Error.
- core/ipc/events.ts: attachIpcEvents(store) — subscribes 'log',
  'render_progress', 'model_download_progress' exactly once (called
  from main.tsx), writing into render.slice (new `progress: number`
  field) and captions.slice (new `modelDownload: {name,pct}|null`
  field) instead of each component re-subscribing.

## Minimal migration (StepExport.tsx, StepTranscript.tsx)

- StepExport: dropped its local listen('log'/'render_progress') effect
  and `progress` useState (now store.progress); write_ass/
  render_audiogram/open_folder calls now go through `ipc.*`.
  render_audiogram's params object is now checked against the actual
  RenderJobDto type — which caught a real pre-existing bug: the old
  code sent an `intro_duration: 3` field RenderJobDto never declared,
  silently dropped by serde's default unknown-field tolerance. Removed
  it; it never did anything.
- StepTranscript: dropped its local listen('model_download_progress')
  effect and downloadingModel/downloadPct useState — both now derived
  from store.modelDownload, so every JSX site that reads them is
  unchanged. startDownload() no longer optimistically sets local state
  before the first real event; confirmed against model_repo.rs that
  the backend's polling thread emits the first (0%) progress event
  within ~400ms of the command starting, so there's no meaningful gap.
- transcribe_audio/write_srt/download_model/list_models calls in both
  files are untouched, per the plan — Phase 2/3 territory.

## Verification

tsc clean, eslint 0 errors (15 warnings, down from 17 — two sources
of the @tauri-apps/api/core warning are gone), full vitest suite green
(16 files/52 tests), cargo check clean. `npm run dev` boots the real
window cleanly with `specta_builder.invoke_handler()` replacing
`tauri::generate_handler!` as the sole dispatcher for all 10 commands
— compiling at all is strong evidence nothing got dropped from
collect_commands!. Did not click through a live MP4 export end-to-end
in this session (same tooling ceiling noted in T6/T7 — no access to
drive the native window or a real file-open dialog here); leaning
instead on: the bindings generator proving RenderJobDto's shape
matches what Rust actually deserializes (a stronger guarantee than
before, when a JS object literal could silently drift from the Rust
struct), and the application-layer render logic itself being
completely untouched — only the IPC plumbing calling into it changed.

+++ apps/desktop/src-tauri/src/domain/entities/model_spec.rs
@@ -1,4 +1,5 @@
 use serde::Serialize;
+use specta::Type;
 
 /// Static catalogue of known Whisper models — lives in the domain because
 /// these are business rules (model names, sizes, quality descriptions).
@@ -15,16 +16,19 @@ pub const MODELS: &[ModelSpec] = &[
 pub struct ModelSpec {
     pub name: &'static str,
     pub label: &'static str,
-    pub size_mb: u64,
+    // u32 (not u64): specta's TS exporter rejects 64-bit ints by default since
+    // JS `number` can't represent them precisely. Model sizes-in-MB never get
+    // remotely close to u32::MAX (~4.29 billion), so no real range is lost.
+    pub size_mb: u32,
     pub note: &'static str,
 }
 
 /// Runtime representation including whether the model is downloaded — sent to the frontend.
-#[derive(Debug, Clone, Serialize)]
+#[derive(Debug, Clone, Serialize, Type)]
 pub struct ModelInfo {
     pub name: String,
     pub label: String,
-    pub size_mb: u64,
+    pub size_mb: u32,
     pub note: String,
     pub downloaded: bool,
 }
